Improved OpenVR connection handling
OpenVR connection is now automatically kept alive even after an error (similar to Oculus). Original implementation was a bit half-assed and the app went into a state where user had to manually restart the tracking. Reasoning behind this was to avoid a memory leak resulting from multiple calls to OpenVR.Init().

As a safety feature for marginal situations where the user lets SteamVR run for a long period of time without the headset connected (or other reason for initialization to fail repeatedly) I put in an automatic restart after 300 init-attempts to flush the memory build-up.